change uninstall.php file

+++ uninstall.php
@@ -1,3 +1,5 @@
 <?php 
-define("WP_UNINSTALL_PLUGIN") || exit;
+if (!defined('WP_UNINSTALL_PLUGIN')) {
+    exit;
+}
 delete_option( "bubble-contact-option" );
